+++ capi/extra_simulator/simulator.h
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 11/10/2022
+// Last Modified: 29/10/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -16,7 +16,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-EXPORT_AUTD void AUTDExtraSimulator(IN const char* settings_path, IN uint16_t port, IN const char* ip, IN bool vsync, IN int32_t gpu_idx);
+EXPORT_AUTD void AUTDExtraSimulator(IN const char* settings_path, IN bool vsync, IN int32_t gpu_idx);
 #ifdef __cplusplus
 }
 #endif