+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Rename `AUTDServer` to `TwinCATAUTDServer`
 - Fix [#36](https://github.com/shinolab/autd3/issues/36)
 - Add log settings in SOEM Link CAPI
+- Remove `port` and `ip` settings from Simulator
 
 # 2.4.5